+++ src/VaultSharp/V1/SecretsEngines/GoogleCloud/GoogleCloudSecretsEngineProvider.cs
@@ -18,7 +18,7 @@ namespace VaultSharp.V1.SecretsEngines.GoogleCloud
         {
             Checker.NotNull(roleset, "roleset");
 
-            return await _polymath.MakeVaultApiRequest<Secret<GoogleCloudOAuth2Token>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.GoogleCloud, "/token/" + roleset.Trim('/'), HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+            return await _polymath.MakeVaultApiRequest<Secret<GoogleCloudOAuth2Token>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.GoogleCloud, "/roleset/" + roleset.Trim('/') + "/token", HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
         }
 
         public async Task<Secret<GoogleCloudServiceAccountKey>> GetServiceAccountKeyAsync(string roleset, ServiceAccountKeyAlgorithm keyAlgorithm = ServiceAccountKeyAlgorithm.KEY_ALG_RSA_2048, ServiceAccountPrivateKeyType privateKeyType = ServiceAccountPrivateKeyType.TYPE_GOOGLE_CREDENTIALS_FILE, string timeToLive = "", string mountPoint = null, string wrapTimeToLive = null)
@@ -32,7 +32,7 @@ namespace VaultSharp.V1.SecretsEngines.GoogleCloud
                 ttl = timeToLive
             };
 
-            return await _polymath.MakeVaultApiRequest<Secret<GoogleCloudServiceAccountKey>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.GoogleCloud, "/key/" + roleset.Trim('/'), HttpMethod.Post, requestData, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+            return await _polymath.MakeVaultApiRequest<Secret<GoogleCloudServiceAccountKey>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.GoogleCloud, "/roleset/" + roleset.Trim('/') + "/key", HttpMethod.Post, requestData, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
         }
     }
 }