+++ test/VaultSharp.Samples/Program.cs
@@ -22,7 +22,7 @@ namespace VaultSharp.Samples
 {
     class Program
     {
-        private const string ExpectedVaultVersion = "1.8.0";
+        private const string ExpectedVaultVersion = "1.8.12";
 
         private static IVaultClient _unauthenticatedVaultClient;
         private static IVaultClient _authenticatedVaultClient;